Update CMake Version in Basekit Dockerfile
The 2022 Intel C++ (Nextgen) compiler requires CMake 3.22.2 to work with Visual Studio. For this reason, the minimum required in the CMakeLists.txt might be 3.22.2 for some users, and this exact same CMakeLists.txt will fail when using this Docker container.

The solution proposed is to update the CMake version inside of the Docker container.